feat: provide navigation task to blockerFn
Currently when using the useBlocker hook there is no way to acertain
where the navigation being blocked is headed.

For my usage of blocking this is necceary to allow navigation within a
selected bunch of nested routes, but not allowing the user the exit the
base route. Thought this could be useful feature to have.